+++ project/libs/environment/src/lib/environment.spec.ts
@@ -4,7 +4,7 @@ describe('environment', () => {
 	it('should get current environment', () => {
 		expect(getEnv()).toEqual(<Env>{
 			CERTBOT_EMAIL: '',
-			DRY_RUN: 'Y',
+			DRY_RUN: 'N',
 			NODE_ENV: 'test'
 		});
 	});