+++ project/libs/util/src/index.ts
@@ -1,3 +1,2 @@
 export * from './lib/file';
-export * from './lib/shell';
 export * from './lib/timer';